requirements.txt
PyGithub==1.55
requests==2.31.0
python-dotenv==1.0.0